update tuto_11, tuto12, tuto_13, tuto_13 shows a constant field controller
