+++ SOFA_playground/BeamApater_tuto/tuto_11.py
@@ -1,6 +1,7 @@
 # This file adds collision model into the actuator and add a env. Meanwhile constrains catheter to 2D.
 # now Catheter contains two sections, one is stiff and the other is softer. The parameters of Catheter 
-# use SOFA example parameters. In the unit of mm, kg
+# use SOFA example parameters. In the unit of mm, kg.
+# Here is the example edited by '3instruments_collision.scn'
 # Author: JunAng Wang
 # Contact: wangjunang94@gmail.com
 from stlib3.scene import MainHeader
@@ -9,16 +10,14 @@ import numpy as np
 def createScene(rootNode):
 
     # Header
-    rootNode.addObject('FreeMotionAnimationLoop')
     rootNode.addObject('DefaultVisualManagerLoop')
+    rootNode.addObject('FreeMotionAnimationLoop')
     rootNode.addObject('LCPConstraintSolver',mu=0.1,tolerance=1e-4,maxIt=1000,build_lcp=False)
     rootNode.addObject('CollisionPipeline',draw=0,depth=6,verbose='1')
-    rootNode.addObject('ParallelBruteForceBroadPhase',name='N2')
-    rootNode.addObject('ParallelBVHNarrowPhase')
+    rootNode.addObject('BruteForceBroadPhase',name='N2')
+    rootNode.addObject('BVHNarrowPhase')
     rootNode.addObject('LocalMinDistance',contactDistance=1,alarmDistance=2,name='localmindistance',angleCone=0.5, coneFactor=0.5)
     rootNode.addObject('CollisionResponse',name='Response',response='FrictionContactConstraint')
-    # rootNode.addObject('DefaultAnimationLoop')
-    # rootNode.addObject('DefaultVisualManagerLoop')
     
     # setting VisualStyle, pluginList, and 3D frame coordinate
     rootNode.addObject('VisualStyle', displayFlags=display)
@@ -27,130 +26,109 @@ def createScene(rootNode):
     rootNode.addObject('InteractiveCamera', name='Camera', position= [-50,0,0], lookAt=[0,0,0])
 
     rootNode.gravity = [0.0, 0.0, 0.0] # mm/s^2
-    radius = 1 #mm
-    inner_radius = 0 
-    StraightLength_stiff = 60 #mm
-    StraightLength_soft = 40 #mm
-    spireDiameter = 4000
-    massDensity = 1550e-9 #kg/mm^3
-    youngModulus_stiff = 10000#N/mm^2 = 10000e6Pa (stiff)
-    youngModulus_soft = 10000
-    poissonRatio = 0.3
-    nbsection_soft = 10
-    nbsection_stiff = 40
-
-    
+        
     #---------------------------------------------------------------------------
     # setting catheter topology mesh and wireRestShape
-    GuideCatheter = rootNode.addChild('GuideCatheter')
+    topoLines = rootNode.addChild('topoLines_cath')
     # create straight Section, 60 sections, if want to import from external files using RodMeshSection
-    GuideCatheter.addObject('RodStraightSection', name ='StraightSection_stiff',
-                         youngModulus= youngModulus_stiff, poissonRatio= poissonRatio, radius = radius, innerRadius= inner_radius, massDensity=massDensity, nbEdgesCollis = nbsection_stiff, nbEdgesVisu = 160, length = StraightLength_stiff)
-    GuideCatheter.addObject('RodSpireSection', name ='StraightSection_soft',
-                         youngModulus= youngModulus_soft, poissonRatio= poissonRatio, radius = radius, innerRadius =inner_radius, massDensity=massDensity, nbEdgesCollis = nbsection_soft, nbEdgesVisu = 40, length = StraightLength_soft, spireDiameter=spireDiameter, spireHeight=0)
+    topoLines.addObject('RodStraightSection', name ='StraightSection',
+                         youngModulus= 10000, poissonRatio = 0.3, nbEdgesCollis = 40, nbEdgesVisu = 220, length = 600)
+    topoLines.addObject('RodSpireSection', name ='SpireSection',
+                         youngModulus= 10000, poissonRatio = 0.3,nbEdgesCollis = 10, nbEdgesVisu = 80, length = 400, spireDiameter=4000, spireHeight=0)
     
 
-    GuideCatheter.addObject('WireRestShape', name='GC_RestShape',  
-                                 printLog=True, template='Rigid3d', wireMaterials = '@StraightSection_stiff @StraightSection_soft')
-    GuideCatheter.addObject('EdgeSetTopologyContainer', name='meshLinesBeam')
-    GuideCatheter.addObject('EdgeSetTopologyModifier', name='Modifier')
-    GuideCatheter.addObject('EdgeSetGeometryAlgorithms', name='GeomAlgo', template='Rigid3d')
-    GuideCatheter.addObject('MechanicalObject', name='dofTopo2', template='Rigid3d')
+    topoLines.addObject('WireRestShape', name='catheterRestShape',  
+                                template='Rigid3d', wireMaterials = '@StraightSection @SpireSection')
+    topoLines.addObject('EdgeSetTopologyContainer', name='meshLinesBeam')
+    topoLines.addObject('EdgeSetTopologyModifier', name='Modifier')
+    topoLines.addObject('EdgeSetGeometryAlgorithms', name='GeomAlgo', template='Rigid3d')
+    topoLines.addObject('MechanicalObject', name='dofTopo1', template='Rigid3d')
     
 
 
     #-----------------------------------------------------------------------
     # mechanical model of Catheter
-    BeamModel = rootNode.addChild('BeamModel')
+    InstrumentCombined = rootNode.addChild('InstrumentCombined')
     
-    BeamModel.addObject('EulerImplicitSolver', rayleighStiffness=0.2, printLog=False, rayleighMass=0.1)
-    BeamModel.addObject('BTDLinearSolver', verbose=False)
-    BeamModel.addObject('RegularGridTopology', name='MeshLines', drawEdges=False, 
-                                    nx=nbsection_soft+nbsection_stiff+1, ny=1, nz=1,
-                                    xmax=1, xmin=0.0, ymin=0, ymax=0, zmax=1, zmin=1,
+    InstrumentCombined.addObject('EulerImplicitSolver', rayleighStiffness=0.2, printLog=False, rayleighMass=0.1)
+    InstrumentCombined.addObject('BTDLinearSolver',subpartSolve=False, verification=False, verbose=False)
+    InstrumentCombined.addObject('RegularGridTopology', name='MeshLines', 
+                                    nx=180, ny=1, nz=1,
+                                    xmax=1.0, xmin=0.0, ymin=0, ymax=0, zmax=1, zmin=1,
                                     p0=[0,0,0])
-    DOFs = BeamModel.addObject('MechanicalObject', showIndices=False, name='Instrument_DOFs', template='Rigid3d')
+    InstrumentCombined.addObject('MechanicalObject', name='DOFs', template='Rigid3d', ry=-90)
     # FEM method (BeamInterpolation)
     # plug the catheter RestShape into WireBeamInterpolation 
-    BeamModel.addObject('WireBeamInterpolation', name='BeamInterpolation', WireRestShape='@../GuideCatheter/GC_RestShape', radius=radius, printLog=False)
+    InstrumentCombined.addObject('WireBeamInterpolation', name='InterpolCatheter', WireRestShape='@../topoLines_cath/catheterRestShape', radius=1, printLog=False)
     # compute internal forces
     # computeMass if false, only compute the stiff elastic model
     # massDensity: Density of the mass
     # shearStressComputation: if false, suppress the shear stress in the computation
     # reinforceLength: a separation computation for the error in elongation is performed
-    BeamModel.addObject('AdaptiveBeamForceFieldAndMass', name='BeamForceField', massDensity=massDensity, interpolation='@BeamInterpolation', computeMass=True, reinforceLength= False, shearStressComputation= False)
+    InstrumentCombined.addObject('AdaptiveBeamForceFieldAndMass', name='CatheterForceField', massDensity= 0.00000155, interpolation='@InterpolCatheter')
 
     #Deployment Controller
-    BeamModel.addObject('InterventionalRadiologyController', name='DeployController', template='Rigid3d', instruments='BeamInterpolation', 
-                                    startingPos=[0,0,0,0,0,0,1], xtip=[1, 0, 0], printLog=False, 
-                                    rotationInstrument=[0, 0, 0], step=0.1, speed=1, 
+    InstrumentCombined.addObject('InterventionalRadiologyController', name='IRController', template='Rigid3d', instruments='InterpolCatheter', 
+                                    startingPos=[0,0,0,0,-0.7071068,0,0.7071068], xtip=[1, 0, 0], printLog=False, 
+                                    rotationInstrument=[0, 0, 0], step=3, speed=2, 
                                     listening=True, controlledInstrument=0)
-    BeamModel.addObject('LinearSolverConstraintCorrection', wire_optimization=True)
+    InstrumentCombined.addObject('LinearSolverConstraintCorrection', wire_optimization=True)
 
-    BeamModel.addObject('FixedConstraint', indices=0, name='FixedConstraint')
+    InstrumentCombined.addObject('FixedConstraint', indices=0, name='FixedConstraint')
     # Add constraint box
     # box = [0,0,-50,120,6,70]
     # BeamModel.addObject('BoxROI',name='BoxROI',box = box, drawBoxes=True, doUpdate=False)
-    BeamModel.addObject('RestShapeSpringsForceField', name="RestSPForceField", points='@DeployController.indexFirstNode', angularStiffness=1e8, stiffness=1e8)
+    InstrumentCombined.addObject('RestShapeSpringsForceField', name="RestSPForceField", points='@IRController.indexFirstNode', angularStiffness=1e8, stiffness=1e8)
     # BeamModel.addObject('PartialFixedConstraint', indices = np.arange(nbsection_stiff + nbsection_soft), fixedDirections = [0,0,1,0,0,0])
     
-    #--------------------------------------------------------------------
-    # Visual model: using WireRestShape mesh
-    VisualCatheter = BeamModel.addChild('VisualCatheter')
-    VisualCatheter.addObject('MechanicalObject', name= 'visu_DOFs',template='Vec3d')
-    
-    VisualCatheter.addObject('QuadSetTopologyContainer', name = "visu_container")
-    VisualCatheter.addObject('QuadSetTopologyModifier')
-    VisualCatheter.addObject('QuadSetGeometryAlgorithms')
-    # Mapping WireRestShape mesh to topology container
-    VisualCatheter.addObject('Edge2QuadTopologicalMapping', nbPointsOnEachCircle=10,
-                             radius=radius, input='@../../GuideCatheter/meshLinesBeam', output='@visu_container')
-    # Mapping mesh to Beam Mechanics
-    VisualCatheter.addObject('AdaptiveBeamMapping', input="@../Instrument_DOFs", output="@visu_DOFs", interpolation='@../BeamInterpolation')
-    
-    # Visual beam by OglModel
-    VisuOgl = VisualCatheter.addChild('VisuOgl')
-    VisuOgl.addObject("OglModel", name="visual", color=[0.5,1,0.5], quads="@../visu_container.quads")
-    VisuOgl.addObject('IdentityMapping', input="@../visu_DOFs",output='@visual')
-
     #-----------------------------------------------------------------------
     # Collision model
-    CollisionCatheter = BeamModel.addChild('CollisionCatheter')
-    CollisionCatheter.addObject('MechanicalObject', name='colli_DOFs',template='Vec3d')
+    Collis = InstrumentCombined.addChild('Collis')
+    Collis.activated = True
     # if use 1D line mesh, one can extend to 2D surf mesh
     # if one needs 3D mesh, one has to import stl 3D files or deploys CylinderGridTopology
 
-    # CollisionCatheter.addObject('QuadSetTopologyContainer',name = 'colli_container')
-    # CollisionCatheter.addObject('QuadSetTopologyModifier')
-    # CollisionCatheter.addObject('QuadSetGeometryAlgorithms',showPointIndices=False, showEdgeIndices=False, drawEdges=False)
-    # CollisionCatheter.addObject('Edge2QuadTopologicalMapping',nbPointsOnEachCircle=10,
-    #                             radius=radius, input='@../../GuideCatheter/meshLinesBeam',
-    #                             output='@colli_container')
-
-    # CollisionCatheter.activated = True
-    CollisionCatheter.addObject('EdgeSetTopologyContainer',name='collisEdgeSet')
-    CollisionCatheter.addObject('EdgeSetTopologyModifier',name='colliseEdgeModifier')
-
-    # CylinderGridTopology, nz is the longitudinal direction discretization
-    # CollisionCatheter.addObject('CylinderGridTopology', name ='Cylinder_container', axis= [1,0,0], center = [0,0,0], length=length, radius=radius, nx=5, ny=5, nz=6)
+    Collis.addObject('EdgeSetTopologyContainer',name='collisEdgeSet')
+    Collis.addObject('EdgeSetTopologyModifier',name='colliseEdgeModifier')
+    Collis.addObject('MechanicalObject', name='collisionDOFs')
     
-    CollisionCatheter.addObject('PointCollisionModel')
-    CollisionCatheter.addObject('LineCollisionModel')
-    # CollisionCatheter.addObject('TriangleCollisionModel')
-    # CollisionCatheter.addObject('TetrahedronCollisionModel')
-    CollisionCatheter.addObject('MultiAdaptiveBeamMapping',controller='../DeployController',useCurvAbs=True, printLog=False,name='collisMap')
+    Collis.addObject('MultiAdaptiveBeamMapping',controller='../IRController',useCurvAbs=True, printLog=False,name='collisMap')
+    Collis.addObject('PointCollisionModel', proximity = 0.0, group ='1')
+    Collis.addObject('LineCollisionModel', proximity=0.0, group = '1')
 
+    #--------------------------------------------------------------------
+    # Visual model: using WireRestShape mesh
+    VisuCatheter = InstrumentCombined.addChild('VisuCatheter')
+    VisuCatheter.activated = True
+    
+    VisuCatheter.addObject('MechanicalObject', name= 'Quads')
+    VisuCatheter.addObject('QuadSetTopologyContainer', name = "ContainerCath")
+    VisuCatheter.addObject('QuadSetTopologyModifier')
+    VisuCatheter.addObject('QuadSetGeometryAlgorithms', template='Vec3d')
+    # Mapping WireRestShape mesh to topology container
+    VisuCatheter.addObject('Edge2QuadTopologicalMapping', nbPointsOnEachCircle=10,
+                             radius=2, input='@../../topoLines_cath/meshLinesBeam', output='@ContainerCath', flipNormals= '1')
+    # Mapping mesh to Beam Mechanics
+    VisuCatheter.addObject('AdaptiveBeamMapping', name='VisuMapCath', useCurvAbs=True, input="@../DOFs", output="@Quads", interpolation='@../InterpolCatheter', isMechanical=False)
+
+    
+    # Visual beam by OglModel
+    VisuOgl = VisuCatheter.addChild('VisuOgl')
+    VisuOgl.addObject("OglModel", name="Visual", color=[0.7,0.7,0.7], quads="@../ContainerCath.quads")
+    VisuOgl.addObject('IdentityMapping', input="@../Quads",output='@Visual')
+
+   
     #---------------------------------------------------------------------
-    # Env = rootNode.addChild('Env')
-    # Env.addObject('MeshOBJLoader', name = 'env_mesh', filename= 'phantom.obj', translation=[0, 0, 0], rotation= [0.0, 180.0, 0.0],triangulate=True, scale = 1)
-    # # Env.addObject('MeshSTLLoader', name = 'env_mesh', filename= 'flat_model_circles.stl', translation=[0, 0, -3.0], rotation= [0.0, 0.0, 0.0],triangulate=True, scale = 1)
-    # Env.addObject('MeshTopology', position = '@env_mesh.position', triangles= '@env_mesh.triangles', drawTriangles=False)
-    # Env.addObject('MechanicalObject', name = 'env_DOFs')
+    CollisionModel = rootNode.addChild('CollisionModel')
+    CollisionModel.addObject('MeshOBJLoader', name = 'meshLoader', filename= 'phantom.obj',triangulate=True, flipNormals='1')
+    # Env.addObject('MeshSTLLoader', name = 'env_mesh', filename= 'flat_model_circles.stl', translation=[0, 0, -3.0], rotation= [0.0, 0.0, 0.0],triangulate=True, scale = 1)
+    CollisionModel.addObject('MeshTopology', position = '@meshLoader.position', triangles= '@meshLoader.triangles', drawTriangles=False)
+    CollisionModel.addObject('MechanicalObject', name = 'DOFs1', position=[0,0,400], scale=3, ry=90)
     
-    # Env.addObject('PointCollisionModel', moving=False, simulated = False)
-    # Env.addObject('TriangleCollisionModel', moving=False, simulated = False)
-    # Env.addObject('LineCollisionModel', moving= False, simulated = False)
-    # Env.addObject('OglModel', name='visu_env', src='@env_mesh', color=[1,0,0,0.3])
+    CollisionModel.addObject('PointCollisionModel', moving=False, simulated = False)
+    CollisionModel.addObject('TriangleCollisionModel', moving=False, simulated = False)
+    CollisionModel.addObject('LineCollisionModel', moving= False, simulated = False)
+    CollisionModel.addObject('OglModel', name='Visual', src='@meshLoader', color=[1,0,0,0.3], scale=3, ry=90)
 
     return rootNode
