+++ SOFA_playground/BeamApater_tuto/tuto_12_2D.py.log
@@ -1,0 +1,6 @@
+
+--- NEW SESSION: 15/01/2024 12:02:03 ---
+//Response.response = FrictionContactConstraint
+
+--- NEW SESSION: 15/01/2024 12:07:31 ---
+//Response.response = StickContactForceField
