+++ SOFA_playground/BeamApater_tuto/tuto_13_2D.py.qglviewer.view
@@ -1,0 +1,2 @@
+75.4789 289.995 1830.66
+0.0783897 0.0733374 -0.00584754 -0.994204
